requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.3
django-allauth==0.57.0
djangorestframework==3.14.0
drf-yasg==1.21.7
GDAL==3.6.4
idna==3.4
inflection==0.5.1
numpy==1.25.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.4
